% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_flowset_samples.R
\name{rename_flowset_samples}
\alias{rename_flowset_samples}
\title{Rename the "name" variable within a flowset}
\usage{
rename_flowset_samples(flo_set, new_names)
}
\arguments{
\item{flo_set}{A flow set produced using \code{flowCore::read.flowSet()}}

\item{new_names}{A character vector of new sample names, of the same length as your flow set}
}
\value{
A flow set with updated names.
}
\description{
A minor annoyance is specifying clean names for plotting many samples within a flowset. This simplifies the process.
}
